requirements.txt
openai==1.12.0
langchain==0.1.10
langchain-openai==0.0.5
chromadb==0.4.22
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
streamlit==1.31.0
folium==0.15.1
streamlit-folium==0.18.0
supabase==2.3.4
requests==2.31.0
python-dotenv==1.0.1
geopy==2.4.1
